音频转srt/requirements.txt
PySide6>=6.4.0
groq>=0.4.0
deepgram-sdk>=2.11.0
httpx>=0.24.0
httpx-socks>=0.7.0
